Heavily modified to integrate with Ozbotics Value system. Note: This version is not backwards compatible with original ArduinoMenu

+++ menuLCDs.h
@@ -23,16 +23,16 @@ www.r-site.net
   class menuLCD:public menuOut {
     public:
     LCD& lcd;
-    menuLCD(LCD& lcd,int x=16,int y=1):lcd(lcd),menuOut(x,y) {}
-    virtual void clearLine(int ln) {
+    menuLCD(LCD& lcd,int8_t x=16,int8_t y=1):lcd(lcd),menuOut(x,y) {}
+    virtual void clearLine(int8_t ln) {
 			setCursor(0,ln);
-    	for(int n=0;n<maxX;n++) print(' ');
+    	for(int8_t n=0;n<maxX;n++) print(' ');
 			setCursor(0,ln);
     }//clear current line
     virtual void clear() {lcd.clear();}
-    virtual void setCursor(int x,int y) {lcd.setCursor(x*resX,y*resY);}
+    virtual void setCursor(int8_t x,int8_t y) {lcd.setCursor(x*resX,y*resY);}
     virtual size_t write(uint8_t ch) {return lcd.write(ch);}
-		virtual void printPrompt(prompt &o,bool selected,int idx,int posY,int width) {
+		virtual void printPrompt(prompt &o,bool selected,int8_t idx,int8_t posY,int8_t width) {
 			clearLine(posY);
 			print(selected?(o.enabled?menu::enabledCursor:menu::disabledCursor):' ');
 			o.printTo(*this);
@@ -41,7 +41,7 @@ www.r-site.net
 			if (m.sel-top>=maxY) top=m.sel-maxY+1;//selected option outside device (bottom)
 			else if (m.sel<top) top=m.sel;//selected option outside device (top)
 			if (drawn!=&m||top!=lastTop) clear();
-			int i=top;for(;i<m.sz;i++) {
+			int8_t i=top;for(;i<m.sz;i++) {
 			  if(i-top>=maxY) break;
 			  if (needRedraw(m,i))
 			  	printPrompt(*m.data[i],i==m.sel,i+1,i-top,m.width);
